rename: detail 페이지 nav 컴포넌트 폴더명 nav로 변경
