+++ src/pages/detail/nav/NoticeContent.tsx
