+++ src/pages/detail/nav/ScheduleContent.tsx
